Improve tracing
This adds tracing of cancellation contexts, and documents and cleans up
the API a bit.

When reading events, instead of the consumer providing one function per
event, we now use a single function taking a variant. This makes the API
easier to use and lets the caller decide whether they want to handle all
events (checked by the compiler) or just some of them.

Co-authored-by: Lucas Pluvinage <lucas@tarides.com>